Created a revised solution with functions
The main failing in my initial implementation was the data analyst's
job to create the script to query the FHIR server. In this improved
solution, I wrote a separate function to get patients from that server
based on very specific ages and conditions. This can be extended in
future versions.
This also separates out the functionality of defining the cohorts,
getting the data and writing the outreach emails.